dir improvements: move `snapshotsync` from `turbo` to `db`
