+++ db/snapshotsync/freezeblocks/beacon_block_reader.go
