+++ db/snapshotsync/freezeblocks/block_txnum_cache.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/erigontech/erigon/turbo/snapshotsync"
+	"github.com/erigontech/erigon/db/snapshotsync"
 )
 
 const E2StepSize = 1_000
