+++ polygon/heimdall/snapshot_store.go
@@ -13,9 +13,9 @@ import (
 	"github.com/erigontech/erigon-lib/common/generics"
 	"github.com/erigontech/erigon-lib/log/v3"
 	"github.com/erigontech/erigon/db/kv"
+	"github.com/erigontech/erigon/db/snapshotsync"
 	"github.com/erigontech/erigon/db/snaptype"
 	"github.com/erigontech/erigon/execution/chain"
-	"github.com/erigontech/erigon/turbo/snapshotsync"
 )
 
 var (
