+++ polygon/heimdall/snapshots.go
@@ -18,8 +18,8 @@ package heimdall
 
 import (
 	"github.com/erigontech/erigon-lib/log/v3"
+	"github.com/erigontech/erigon/db/snapshotsync"
 	"github.com/erigontech/erigon/eth/ethconfig"
-	"github.com/erigontech/erigon/turbo/snapshotsync"
 )
 
 // Bor Events
